pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]

name = "fastapi-template"
version = "0.1.0"
description = "A production-ready FastAPI starter template"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.1",
    "sqlalchemy[asyncio]>=2.0.32",
    "asyncpg>=0.29.0",
    "redis>=5.0.8",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.1",
    "email-validator>=2.2.0",
    "alembic>=1.17.2",
    "PyYAML>=6.0.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.2",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.2",
    "aiosqlite>=0.20.0",
    "greenlet>=3.0.3",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project.scripts]
dev = "app.cli:dev"
start = "app.cli:start"
test = "app.cli:test"
migrate = "app.cli:migrate"
makemigrations = "app.cli:makemigrations"
